flutter-engine: gn-utils.inc: fix arch translation of gn_clang_triple_prefix
The value of GN_TARGET_ARCH_NAME_aarch64 is "arm64".
However, the match expression of gn_clang_triple_prefix is
"aarch64.*".
Then, the target triple for "aarch64" could not work and
the wrong configuration, "armv7-unknown-linux-gnueabihf" is used.

So, I fix them to let it choise "aarch64-unknown-linux-gnu" if aarch64
building is done.

Signed-off-by: Kazuyoshi Akiyama <akiyama@nds-osk.co.jp>